support-copilot/requirements.txt
langgraph 
langchain 
langchain-openai 
langsmith 
python-dotenv 
faiss-cpu